requirements.txt
beautifulsoup4>=4.12.0
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
openai>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0